Bound DeepSeek V4 decode: materialize the compressor carry buffers
Decode retained ~0.52GB per token that no Python-side accounting could
see. The KV rings, compressed stores and weight cache were all flat, and
a gc census of every live mx.array stayed at 1.39GB while active memory
climbed to 4.32GB over twelve tokens -- the retained buffers had no
Python wrapper at all.

CompressorState._write rebuilds kv_state by concatenating around the
PREVIOUS kv_state. Left unevaluated that chains one graph link per decode
step per layer, and each link pins the projections that fed it, and
through them the compressor's own [1024, 4096] wkv and wgate matrices:
16.8MB per compressed layer per step, ~0.69GB per token across 41
compressed layers, held for the life of the request. The period-4
sawtooth in the trace was the group-completion slide breaking part of the
chain, and the -1.058GB steps every fourth sweep were the LM head being
evicted and refetched, not a release.

Evaluating the two carry buffers at the end of each write flattens it.
Measured over 24 decode tokens: active memory constant at 1.39GB for
every sweep (was 1.59 -> 4.32GB over twelve), peak Metal 6.07GB ->
2.80GB, same greedy output.

Also stop upcasting the compressor weights to float32 on every step.
Projecting in the weights' own dtype and widening the result computes the
same pooling in float32 without allocating a float32 copy of both
matrices per layer per step. This does not show in end-of-sweep active
memory -- the copies were transient -- but it removes them from peak.

F219 asserts the property rather than the call: retention over 40 steps
must not exceed retention over 4, stepping must leave no pending work,
and emitted values must be unchanged.

+++ runtime/engine.py
@@ -2836,9 +2836,15 @@ class StreamingEngine:
                 states[layer] = state
             cw = w[f"{prefix}.attn.compressor.wkv.weight"]
             cg = w[f"{prefix}.attn.compressor.wgate.weight"]
+            # Project in the weights' own dtype and widen only the RESULT.
+            # Upcasting the weights instead allocated a float32 copy of both
+            # [1024, 4096] matrices on every decode step for every compressed
+            # layer -- about 16.8MB x 41 layers = 0.69GB per token, which is
+            # the sawtooth that remained after the ring fix. Pooling still
+            # happens in float32, which is what the released module widens for.
             pooled = state.step(
-                (x.astype(mx.float32) @ cw.astype(mx.float32).T),
-                (x.astype(mx.float32) @ cg.astype(mx.float32).T),
+                (x.astype(cw.dtype) @ cw.T).astype(mx.float32),
+                (x.astype(cg.dtype) @ cg.T).astype(mx.float32),
                 offset, w[f"{prefix}.attn.compressor.ape"])
             if pooled is not None:
                 pooled = mx.fast.rms_norm(
